requirement.txt
numpy
pillow
scikit-image
fire
matplotlib
opencv-python
numba
easydict
tensorflow >= 2.0
cython
tqdm
pyquaternion
compressai
